requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
chardet==5.2.0
contourpy==1.3.0
cycler==0.12.1
discord-ext-pages==0.0.1
discord.py==2.5.0
et_xmlfile==2.0.0
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
multidict==6.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
zipp==3.21.0
